\name{haploOR-internal}
\title{Internal Functions for haplotypeOddsRatio}
\alias{genhaplopairs}
\alias{hwehaplofreq}
\alias{probonecopy}
\description{
 Internal functions needed for haplotypeOddsRatio.
}
\usage{
genhaplopairs(n)
hwehaplofreq(n, two2k, g1idx, g2idx, g1tbl, hpair, tol=1e-8)
probonecopy(n, hf0, g1idx, g1tbl, hpair)
}
\details{
  These functions are not to be called by the user.
}
\keyword{internal}
