\name{genepi-package}
\alias{genepi-package}
\alias{genepi}
\docType{package}
\title{
  Functions for some genetic epidemiology methods.
}
\description{
  These functions provide code for genetic epidemiology methods
  developed at MSKCC.  They currently include estimating haplotype
  disease risk and two stage designs for GWAS.
}
\details{
  
  \tabular{ll}{
    Package: \tab genepi\cr
    Type: \tab Package\cr
    Version: \tab 1.0.3\cr
    Date: \tab 2023-08-31\cr
    License: \tab GPL version 2 or later\cr
    LazyLoad: \tab yes\cr
  }
  There are two functions \code{haplotypeOddsRatio} and
  \code{twoStagePower} in this package.
  
  Package will be archived and functions added to clinfun package
}
\author{
  Venkatraman E. Seshan
  
  Maintainer: Venkatraman E. Seshan <seshanv@mskcc.org>
}
\references{
  Satagopan JM, Venkatraman ES, Begg CB. (2004) Two-stage designs for
  gene-disease association studies with sample size constraints.
  \emph{Biometrics} 
 
  Venkatraman ES, Mitra N, Begg CB. (2004) A method for evaluating the
  impact of individual haplotypes on disease incidence in molecular
  epidemiology studies. \emph{Stat Appl Genet Mol Biol.} v3:Article27.
}
\keyword{package}
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
